Fixed issues with loading comment mark schema in non collaborative contexts and added codeblock support
