+++ packages/text/src/index.ts
@@ -24,6 +24,7 @@ export * from './nodes'
 export * from './marks/code'
 export * from './marks/colors'
 export * from './marks/noteBase'
+export * from './marks/inlineComment'
 export * from './markdown'
 export * from './markdown/serializer'
 export * from './markdown/parser'
