+++ packages/text/src/kits/server-kit.ts
@@ -37,6 +37,7 @@ import { MermaidExtension, mermaidOptions } from '../nodes/mermaid'
 import TextAlign from '@tiptap/extension-text-align'
 import TextStyle from '@tiptap/extension-text-style'
 import { BackgroundColor, TextColor } from '../marks/colors'
+import { InlineCommentMark } from '../marks/inlineComment'
 
 const headingLevels: Level[] = [1, 2, 3, 4, 5, 6]
 
@@ -84,6 +85,7 @@ export const ServerKit = Extension.create<ServerKitOptions>({
           levels: headingLevels
         }
       }),
+      InlineCommentMark.configure({}),
       CodeBlockExtension.configure(codeBlockOptions),
       CodeExtension.configure(codeOptions),
       MermaidExtension.configure(mermaidOptions),
